Simplify python mdsExceptions implementation
The mdsshr/gen_messages.py script now generates a much
simpler mdsobjects/python/mdsExceptions.py module instead
of building mdsExceptions as a subpackage. This greatly
simplifies things without any need to do special imports
and global inspections which was there in mdsExceptions/__init__.py.
