+++ mdsobjects/python/setup.py
@@ -64,13 +64,12 @@ setup(name=name,
                      pname+'.tests':'./tests',
                      pname+'.widgets':'./widgets',
                      pname+'.wsgi':'./wsgi',
-                     pname+'.mdsExceptions':'./mdsExceptions'},
+                     },
       packages = [pname,
                   pname+'.tdibuiltins',
                   pname+'.tests',
                   pname+'.widgets',
-                  pname+'.wsgi',
-                  pname+'.mdsExceptions'],
+                  pname+'.wsgi'],
       package_data = {'':['doc/*.*','widgets/*.glade','js/*.js','html/*.html','wsgi/*.tbl']},
       include_package_data = True,
       classifiers = [
